Create README.md
TACTICS 🛍️ is a responsive e-commerce platform designed for seamless shopping. Built with the MVC architecture, it uses EJS for dynamic views, Node.js and Express.js for the backend, and MongoDB for data management. TACTICS offers a clean design, secure payments, and fast performance, making it perfect for both users and developers exploring modern web technologies.